bext

+++ 0x06-pointers_arrays_strings/101-print_number.c
@@ -12,9 +12,9 @@ int _putchar(char c);
 
 void print_number(int n)
 {
-	long m;	  /* power of 10 */
-	int c;	  /* boolean check */
-	long num; /* convert int to long */
+	int m;
+	int c;
+	int num;
 
 	num = n;
 	/* negatives */
